third_party/blink/web_tests/external/wpt/css/css-properties-values-api/at-property-expected.txt

This is a testharness.js-based test.
Found 88 tests; 60 PASS, 28 FAIL, 0 TIMEOUT, 0 NOTRUN.
[PASS] Attribute 'syntax' returns expected value for ["<color>"]
[PASS] Attribute 'syntax' returns expected value for ["<color> | none"]
[PASS] Attribute 'syntax' returns expected value for ["<color># | <image> | none"]
[PASS] Attribute 'syntax' returns expected value for ["foo | <length>#"]
[PASS] Attribute 'syntax' returns expected value for ["foo | bar | baz"]
[PASS] Attribute 'syntax' returns expected value for ["notasyntax"]
[PASS] Attribute 'syntax' returns expected value for ["*"]
[PASS] Attribute 'syntax' returns expected value for [" * "]
[PASS] Attribute 'syntax' returns expected value for ["* "]
[PASS] Attribute 'syntax' returns expected value for ["	*	"]
[PASS] Attribute 'syntax' returns expected value for ["red"]
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["rgb(255, 0, 0)"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for [<color>]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for [foo | bar]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["default"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["Default"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["initial"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["Initial"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["inherit"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["Inherit"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["unset"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["Unset"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["revert"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["Revert"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["revert-layer"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["Revert-layer"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["foo bar"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["Foo <length>"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["foo, bar"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["<length> <percentage>"]
  assert_true: expected true got false
[FAIL] Attribute 'syntax' makes the @property rule invalid for ["|<length>"]
  assert_true: expected true got false
[PASS] Attribute 'initial-value' returns expected value for [10px]
[PASS] Attribute 'initial-value' returns expected value for [rgb(1, 2, 3)]
[PASS] Attribute 'initial-value' returns expected value for [red]
[PASS] Attribute 'initial-value' returns expected value for [foo]
[PASS] Attribute 'initial-value' returns expected value for [if(){}]
[FAIL] Attribute 'initial-value' makes the @property rule invalid for [3em]
  assert_true: expected true got false
[FAIL] Attribute 'initial-value' makes the @property rule invalid for [var(--x)]
  assert_true: expected true got false
[PASS] Attribute 'inherits' returns expected value for [true]
[PASS] Attribute 'inherits' returns expected value for [false]
[FAIL] Attribute 'inherits' makes the @property rule invalid for [none]
  assert_true: expected true got false
[FAIL] Attribute 'inherits' makes the @property rule invalid for [0]
  assert_true: expected true got false
[FAIL] Attribute 'inherits' makes the @property rule invalid for [1]
  assert_true: expected true got false
[FAIL] Attribute 'inherits' makes the @property rule invalid for ["true"]
  assert_true: expected true got false
[FAIL] Attribute 'inherits' makes the @property rule invalid for ["false"]
  assert_true: expected true got false
[FAIL] Attribute 'inherits' makes the @property rule invalid for [calc(0)]
  assert_true: expected true got false
[PASS] Invalid property name does not parse [foo]
[PASS] Invalid property name does not parse [-foo]
[PASS] Rule applied [*, if(){}, false]
[PASS] Rule applied [<angle>, 42deg, false]
[PASS] Rule applied [<angle>, 1turn, false]
[PASS] Rule applied [<color>, green, false]
[PASS] Rule applied [<color>, rgb(1, 2, 3), false]
[PASS] Rule applied [<image>, url("http://a/"), false]
[PASS] Rule applied [<integer>, 5, false]
[PASS] Rule applied [<length-percentage>, 10px, false]
[PASS] Rule applied [<length-percentage>, 10%, false]
[PASS] Rule applied [<length-percentage>, calc(10% + 10px), false]
[PASS] Rule applied [<length>, 10px, false]
[PASS] Rule applied [<number>, 2.5, false]
[PASS] Rule applied [<percentage>, 10%, false]
[PASS] Rule applied [<resolution>, 50dppx, false]
[PASS] Rule applied [<resolution>, 96dpi, false]
[PASS] Rule applied [<time>, 10s, false]
[PASS] Rule applied [<time>, 1000ms, false]
[PASS] Rule applied [<transform-function>, rotateX(0deg), false]
[PASS] Rule applied [<transform-list>, rotateX(0deg), false]
[PASS] Rule applied [<transform-list>, rotateX(0deg) translateX(10px), false]
[PASS] Rule applied [<url>, url("http://a/"), false]
[PASS] Rule applied [<color>, tomato, false]
[PASS] Rule applied [<color>, tomato, true]
[PASS] Rule applied for "*", even with no initial value
[PASS] Rule not applied [undefined, green, false]
[PASS] Rule not applied [<color>, undefined, false]
[PASS] Rule not applied [<color>, green, undefined]
[PASS] Rule not applied [<gandalf>, grey, false]
[PASS] Rule not applied [gandalf, grey, false]
[PASS] Rule not applied [<color>, notacolor, false]
[PASS] Rule not applied [<length>, 10em, false]
[PASS] Rule not applied [<transform-function>, translateX(1em), false]
[PASS] Rule not applied [<transform-function>, translateY(1lh), false]
[PASS] Rule not applied [<transform-list>, rotate(10deg) translateX(1em), false]
[PASS] Rule not applied [<transform-list>, rotate(10deg) translateY(1lh), false]
[PASS] Non-inherited properties do not inherit
[PASS] Inherited properties inherit
[PASS] Initial values substituted as computed value
[PASS] Non-universal registration are invalid without an initial value
[PASS] Initial value may be omitted for universal registration
Harness: the test ran to completion.
